--- packages/zevm-contracts/scripts/cross-chain-counter/deploy.ts
@@ -1,10 +1,0 @@
-import hardhat from "hardhat";
-
-const main = async () => {
-  console.log(`Deploying ZetaSwap...`);
-};
-
-main().catch(error => {
-  console.error(error);
-  process.exit(1);
-});
