+++ packages/zevm-contracts/scripts/systemConstants.ts
@@ -1,0 +1,4 @@
+export const SWAP_ADDRESS = "0x43cFaA5cDd103B7733B13cF72e21aB97A2074C84";
+export const TOKEN_ADDRESS = "0x6e78f64edc4c422e7045bcf7562f13bbee75948a";
+export const MATH_ADDRESS = "0x66aa280f3db58231d08835673af286a8aaa80ef5";
+export const VIEWS_ADDRESS = "0x1bfef792f3b0483a6e201c004d9a59de214cb68d";
